feat(emails): add SignupEmail component for user verification
- Introduced a new SignupEmail component to send verification emails to users during the signup process.
- Updated the signup action to dynamically import the SignupEmail component server-side, ensuring it is not bundled for the client.
- Modified the login route to correctly encode the redirect URL for improved navigation.
