+++ app/emails/signup-email.tsx
@@ -1,0 +1,29 @@
+import * as E from '@react-email/components';
+
+export const SignupEmail = ({
+  onboardingUrl,
+  otp,
+}: {
+  onboardingUrl: string;
+  otp: string;
+}) => {
+  return (
+    <E.Html lang="en" dir="ltr">
+      <E.Container>
+        <h1>
+          <E.Text>Welcome to GiftPool!</E.Text>
+        </h1>
+        <p>
+          <E.Text>
+            Here's your verification code: <strong>{otp}</strong>
+          </E.Text>
+        </p>
+        <p>
+          <E.Text>Or click the link to get started:</E.Text>
+        </p>
+        <E.Link href={onboardingUrl}>{onboardingUrl}</E.Link>
+      </E.Container>
+    </E.Html>
+  );
+};
+
