+++ app/routes/_auth+/login.tsx
@@ -174,7 +174,7 @@ const LoginPage = () => {
               <Link
                 to={
                   redirectTo
-                    ? `/signup?${encodeURIComponent(redirectTo)}`
+                    ? `/signup?redirectTo=${encodeURIComponent(redirectTo)}`
                     : '/signup'
                 }
               >
